[SCFToCalyx] [StaticLogicToCalyx] Continue separation conversions.
